requirements.txt
cryptography==3.3.2
aioredis==1.3.1
asgiref==3.4.1
gunicorn==19.7.1
async-timeout==3.0.1
attrs==21.2.0
autobahn==18.10.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
constantly==15.1.0
daphne==3.0.2
Django==3.2.8
django-after-response==0.2.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-simplejwt
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.2
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pytz==2021.3
requests==2.26.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
Twisted==19.7.0
txaio==21.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
zope.interface==5.4.0
asn1crypto==1.4.0
phonenumbers==8.12.21
python-dotenv==0.17.0
twilio==6.57.0
authy==2.2.6